Added village trade station
-fixed issue with atm crash on withdrawal
-widened atm gui to fit full value of long for account balance
-added trade station to village gen
-changed a couple blocks to handle full value of long for coinSum

+++ java/universalcoins/container/ContainerPackager.java
@@ -12,7 +12,7 @@ import universalcoins.tileentity.TilePackager;
 public class ContainerPackager extends Container {
 
 	private TilePackager tEntity;
-	private int lastCoinSum;
+	private long lastCoinSum;
 	private int lastPackageSize;
 	private String lastPackageTarget = "";
 	private boolean lastCardAvailable;
